+++ changelog.md
@@ -2,5 +2,6 @@
 
 ## 20231228
 
+- Exporter: Add export rule to include/exclude template categories
 - Prism: Update library to 1.29.0 - Replace Clipboard.js with browser-native `navigator.clipboard`
 - Template edit screen: Add Preview tab